Add files via upload
Files uploaded are mathmatical problems that are suitable for anyone studying machine learning related topics. The questions contained are generated based on top institutions' test content for research assistant positions, and the difference between the full math problem set and the other one is that the latter one did not contain questions that are related to theories of convergence in probility and convergence in distribution. Visitors are more than welcome to use the content for self-practice, and please do not use the file here for any commercial purpose.